fix lint error by using ul instead of cast

+++ src/PubSubClient.cpp
@@ -373,13 +373,13 @@ bool PubSubClient::handlePacket(uint8_t hdrLen, size_t length) {
                 // - Payload (for QoS > 0): length - (hdrLen + 5 + topicLen) bytes (starts at _buffer[hdrLen + 5 + topicLen])
                 // To get a null reminated 'C' topic string we move the topic 1 byte to the front (overwriting the LSB of the topic lenght)
                 // Guard 1: ensure topic length bytes are readable
-                if ((hdrLen + 2ul >= length) || ((size_t)hdrLen + 2 >= _bufferSize)) {
+                if ((hdrLen + 2ul >= length) || (hdrLen + 2ul >= _bufferSize)) {
                     ERROR_PSC_PRINTF_P("handlePacket(): Packet too short to contain topic length field\n");
                     return false;
                 }
                 uint16_t topicLen = (_buffer[hdrLen + 1] << 8) + _buffer[hdrLen + 2];  // topic length in bytes
                 char* topic = (char*)(_buffer + hdrLen + 3 - 1);                       // set the topic in the LSB of the topic lenght, as we move it there
-                size_t payloadOffset = (size_t)hdrLen + 3 + topicLen;  // payload starts after header and topic (if there is no packet identifier)
+                size_t payloadOffset = hdrLen + 3ul + topicLen;  // payload starts after header and topic (if there is no packet identifier)
                 size_t payloadLen = length - payloadOffset;            // this might change by 2 if we have a QoS 1 or 2 message
                 uint8_t* payload = _buffer + payloadOffset;
